powertag2mqtt/requirements.txt
paho-mqtt
pyhocon
#ruamel.yaml
